Implemented MDA widget Refactor pylab modules for global mmcore and Config access
